[RUST] split test_table.py
